+++ src/Bio/Motions/Types.hs
@@ -37,7 +37,7 @@ data Input = Input
     { inputChainLength :: Int
     , inputLamins :: [Int]
     , inputBinders :: [Int]
-    , inputRadius :: Double
+    , inputRadius :: Int
     , inputNumBinders :: Int
     , inputNumSteps :: Int
     , inputRandList :: [SimRand]
